all models adhere to model_lib_core.py
