+++ models/qwen2.5_vl_3b/qwen2.5_vl_3b_config.json
@@ -282,7 +282,7 @@
   },
   "defaults": {
     "prompt": "Describe this image.",
-    "image": "test_samples/test.jpg",
+    "image": "test_samples/test_image.jpg",
     "dev": "xdma0",
     "cycle_ns": 5.63,
     "max_seq": 512
