+++ models/smolvlm2/smolvlm2_test.py
@@ -28,6 +28,13 @@ from user_dma_core import (
     DRAM_INSTRUCTION_ADDR, INSTRUCTION_REG_REWRITE, MEMCPY_TYPE,
     UnifiedEngine,
 )
+from model_lib_core import (
+    store_weight, store_quantized_weight, load_weight_cache, store_identity_matrix,
+    eltwise_add_core_dram, eltwise_mul_core_dram,
+    rms_norm_core_dram_post_add, layer_norm_core_dram_post_add,
+    quantize_q4_64 as _mlc_quantize_q4_64,
+    quantize_fp4_64 as _mlc_quantize_fp4_64,
+)
 def _load_smolvlm2_config(path: str | None = None) -> dict:
     if path is None:
         path = os.path.join(os.path.dirname(os.path.abspath(__file__)), "smolvlm2_config.json")
@@ -39,51 +46,6 @@ HF_MODEL_REPO = _SMOLVLM2_CFG["paths"]["hf_model_repo"]
 # =============================================================================
 # Helper Methods for SmolVLM2
 # =============================================================================
-def store_weight(ue, tensor, padded_shape=None):
-    """Pad, convert to bf16, DMA to device DRAM. Returns DRAM address."""
-    bf16 = tensor.to(torch.bfloat16)
-    if padded_shape is not None:
-        padded = torch.zeros(padded_shape, dtype=torch.bfloat16)
-        slices = tuple(slice(0, s) for s in bf16.shape)
-        padded[slices] = bf16
-        bf16 = padded
-    bf16 = bf16.contiguous()
-    nbytes = bf16.numel() * 2
-    addr = ue.get_params_dram_addr()
-    ue.dma_write(DMA_DEVICE_H2C, addr, bf16.flatten(), nbytes)
-    ue.allocate_params_dram(nbytes)
-    return addr
-def store_quantized_weight(ue, raw_data):
-    """Store Q4_64 raw GGUF data as scale + packed in DRAM. Returns (scale_addr, data_addr)."""
-    raw_bytes = raw_data.tobytes() if hasattr(raw_data, 'tobytes') else bytes(raw_data)
-    n_blocks = len(raw_bytes) // 34
-    scales_size = n_blocks * 2
-    data_size = n_blocks * 32
-
-    scales_np = np.frombuffer(raw_bytes[:scales_size], dtype=np.uint16).copy()
-    scale_tensor = torch.from_numpy(scales_np).view(torch.bfloat16)
-    scale_addr = ue.get_params_dram_addr()
-    ue.dma_write(DMA_DEVICE_H2C, scale_addr, scale_tensor, scales_size)
-    ue.allocate_params_dram(scales_size)
-
-    data_np = np.frombuffer(raw_bytes[scales_size:scales_size + data_size], dtype=np.uint8).copy()
-    data_tensor = torch.from_numpy(data_np)
-    data_addr = ue.get_params_dram_addr()
-    ue.dma_write(DMA_DEVICE_H2C, data_addr, data_tensor, data_size)
-    ue.allocate_params_dram(data_size)
-
-    return scale_addr, data_addr
-def load_weight_cache(bin_path):
-    """Load bin+json weight file. Returns {tensor_name: raw_numpy_data}."""
-    json_path = bin_path.rsplit('.', 1)[0] + '.json'
-    with open(json_path) as f:
-        manifest = json.load(f)
-    with open(bin_path, 'rb') as f:
-        raw = f.read()
-    cache = {}
-    for name, meta in manifest.items():
-        cache[name] = np.frombuffer(raw[meta['offset']:meta['offset'] + meta['size']], dtype=np.uint8).copy()
-    return cache
 def init_hang_prevention(ue) -> None:
     """Stop stale execution and write HALT to instruction DRAM base."""
     print("[Init] Hang prevention: disabling instruction execution...")
@@ -140,54 +102,6 @@ def accelerator_memory_to_sram_reg(ue, accelerator_dram_address: int, sram_addre
                            stride_jump_bytes=stride_jump_bytes,
                            general_reg_src=general_reg_src)
     ue._inst_id += 1
-def eltwise_add_core_dram(ue, size: int, A_DRAM_ADDR: int, B_DRAM_ADDR: int, OUTPUT_DRAM_ADDR: int) -> int:
-    """OUTPUT = A + B (DRAM)."""
-    bytes_per_element = 2
-    uram_a_addr = 0x00000
-    uram_b_addr = 0x80000
-    chunk_size = min(URAM_NEAR_FULL_ELEMENTS, size)
-
-    for start, take in ue.chunk_ranges(size, chunk_size):
-        offset_bytes = start * bytes_per_element
-        ue.accelerator_memory_to_sram(
-            accelerator_dram_address=A_DRAM_ADDR + offset_bytes,
-            sram_address=uram_a_addr, element_size=take)
-        ue.accelerator_memory_to_sram(
-            accelerator_dram_address=B_DRAM_ADDR + offset_bytes,
-            sram_address=uram_b_addr, element_size=take)
-        ue.eltwise_add_core(
-            vector_A_sram_start_addr=uram_a_addr,
-            vector_B_sram_start_addr=uram_b_addr,
-            vector_C_sram_wb_addr=uram_a_addr, element_size=take)
-        ue.sram_to_accelerator_memory(
-            sram_address=uram_a_addr,
-            accelerator_dram_address=OUTPUT_DRAM_ADDR + offset_bytes,
-            element_size=take)
-    return size
-def eltwise_mul_core_dram(ue, size: int, A_DRAM_ADDR: int, B_DRAM_ADDR: int, OUTPUT_DRAM_ADDR: int) -> int:
-    """OUTPUT = A * B (DRAM)."""
-    bytes_per_element = 2
-    uram_a_addr = 0x00000
-    uram_b_addr = 0x80000
-    chunk_size = min(URAM_NEAR_FULL_ELEMENTS, size)
-
-    for start, take in ue.chunk_ranges(size, chunk_size):
-        offset_bytes = start * bytes_per_element
-        ue.accelerator_memory_to_sram(
-            accelerator_dram_address=A_DRAM_ADDR + offset_bytes,
-            sram_address=uram_a_addr, element_size=take)
-        ue.accelerator_memory_to_sram(
-            accelerator_dram_address=B_DRAM_ADDR + offset_bytes,
-            sram_address=uram_b_addr, element_size=take)
-        ue.eltwise_mul_core(
-            vector_A_sram_start_addr=uram_a_addr,
-            vector_B_sram_start_addr=uram_b_addr,
-            vector_C_sram_wb_addr=uram_a_addr, element_size=take)
-        ue.sram_to_accelerator_memory(
-            sram_address=uram_a_addr,
-            accelerator_dram_address=OUTPUT_DRAM_ADDR + offset_bytes,
-            element_size=take)
-    return size
 def capture_to_raw(ue):
     """Stop capture, extract raw instruction bytes (no halt), clear buffer."""
     ue.stop_capture()
@@ -206,144 +120,6 @@ def generate_halt_raw(ue):
         raw.extend(inst.get_bytes())
     ue.clear_capture_buffer()
     return bytes(raw)
-def rms_norm_core_dram_post_add(ue, M: int, N: int,
-                                 A_DRAM_ADDR: int, B_DRAM_ADDR: int,
-                                 ADDOUTPUT_DRAM_ADDR: int, NORMOUTPUT_DRAM_ADDR: int,
-                                 GAMMA_DRAM_ADDR: int = None) -> int:
-    """rms_norm(A + B) with residual output."""
-    gamma_sram_addr = 0x80000
-    params_sram_addr = gamma_sram_addr
-
-    if GAMMA_DRAM_ADDR is not None:
-        ue.accelerator_memory_to_sram(accelerator_dram_address=GAMMA_DRAM_ADDR,
-                                    sram_address=gamma_sram_addr, element_size=N)
-        params_sram_addr += N * 2
-    else:
-        gamma_sram_addr = None
-
-    vector_A_sram_addr = 0x00000
-    vector_B_sram_addr = params_sram_addr
-    uram_b_remaining_elements = URAM_NEAR_FULL_ELEMENTS - (params_sram_addr - 0x80000) // 2
-    chunk_size = min(URAM_NEAR_FULL_ELEMENTS // N, uram_b_remaining_elements // N, M)
-    assert chunk_size >= 1 and chunk_size <= M
-
-    for i, m_take in ue.chunk_ranges(M, chunk_size):
-        chunk_elements = m_take * N
-        ue.accelerator_memory_to_sram(accelerator_dram_address=A_DRAM_ADDR + i * N * 2,
-                                    sram_address=vector_A_sram_addr, element_size=chunk_elements)
-        ue.accelerator_memory_to_sram(accelerator_dram_address=B_DRAM_ADDR + i * N * 2,
-                                    sram_address=vector_B_sram_addr, element_size=chunk_elements)
-        for j in range(m_take):
-            ue.eltwise_add_core(vector_A_sram_addr + j * N * 2, vector_B_sram_addr + j * N * 2,
-                                vector_A_sram_addr + j * N * 2, N)
-        ue.sram_to_accelerator_memory(sram_address=vector_A_sram_addr,
-                                        accelerator_dram_address=ADDOUTPUT_DRAM_ADDR + i * N * 2,
-                                        element_size=chunk_elements)
-        for j in range(m_take):
-            ue.rms_norm_core(vector_A_sram_addr + j * N * 2, vector_A_sram_addr + j * N * 2,
-                             N, gamma_sram_addr)
-        ue.sram_to_accelerator_memory(sram_address=vector_A_sram_addr,
-                                        accelerator_dram_address=NORMOUTPUT_DRAM_ADDR + i * N * 2,
-                                        element_size=chunk_elements)
-
-    total_flops = M * N + 3 * M * N
-    if gamma_sram_addr is not None:
-        total_flops += M * N
-    return total_flops
-def layer_norm_core_dram(ue, M: int, N: int,
-                         A_DRAM_ADDR: int, OUTPUT_DRAM_ADDR: int,
-                         GAMMA_DRAM_ADDR: int = None, BETA_DRAM_ADDR: int = None) -> int:
-    """OUTPUT = LayerNorm(A)."""
-    zeros_sram = 0x80000
-    zeros_dram = ue.get_params_dram_addr()
-    ue.allocate_params_dram(N * 2)
-    ue.dma_write(DMA_DEVICE_H2C, zeros_dram, torch.zeros(N, dtype=torch.bfloat16), N * 2)
-    ue.accelerator_memory_to_sram(zeros_dram, zeros_sram, N)
-    params_sram = zeros_sram + N * 2
-
-    gamma_sram = None
-    if GAMMA_DRAM_ADDR is not None:
-        gamma_sram = params_sram
-        ue.accelerator_memory_to_sram(GAMMA_DRAM_ADDR, gamma_sram, N)
-        params_sram += N * 2
-    beta_sram = None
-    if BETA_DRAM_ADDR is not None:
-        beta_sram = params_sram
-        ue.accelerator_memory_to_sram(BETA_DRAM_ADDR, beta_sram, N)
-        params_sram += N * 2
-
-    vector_sram = 0x00000
-    uram_b_used = (params_sram - 0x80000) // 2
-    chunk = min(URAM_NEAR_FULL_ELEMENTS // N, (URAM_NEAR_FULL_ELEMENTS - uram_b_used) // N, M)
-    assert chunk >= 1
-
-    for i, m_take in ue.chunk_ranges(M, chunk):
-        ue.accelerator_memory_to_sram(A_DRAM_ADDR + i * N * 2, vector_sram, m_take * N)
-        for j in range(m_take):
-            ue.layer_norm_core(vector_sram + j * N * 2, vector_sram + j * N * 2, N, zeros_sram, gamma_sram, beta_sram)
-        ue.sram_to_accelerator_memory(vector_sram, OUTPUT_DRAM_ADDR + i * N * 2, m_take * N)
-    return 5 * M * N + (M * N if gamma_sram else 0) + (M * N if beta_sram else 0)
-def layer_norm_core_dram_post_add(ue, M: int, N: int,
-                                   A_DRAM_ADDR: int, B_DRAM_ADDR: int,
-                                   ADDOUTPUT_DRAM_ADDR: int, NORMOUTPUT_DRAM_ADDR: int,
-                                   GAMMA_DRAM_ADDR: int = None, BETA_DRAM_ADDR: int = None) -> int:
-    """layer_norm(A + B) with residual output."""
-    zeros_sram_addr = 0x80000
-
-    zeros_dram_addr = ue.get_params_dram_addr()
-    ue.allocate_params_dram(N * 2)
-    ue.dma_write(DMA_DEVICE_H2C, zeros_dram_addr, torch.zeros(N, dtype=torch.bfloat16), N * 2)
-
-    ue.accelerator_memory_to_sram(accelerator_dram_address=zeros_dram_addr,
-                                    sram_address=zeros_sram_addr, element_size=N)
-    params_sram_addr = zeros_sram_addr + N * 2
-
-    gamma_sram_addr = None
-    beta_sram_addr = None
-
-    if GAMMA_DRAM_ADDR is not None:
-        gamma_sram_addr = params_sram_addr
-        ue.accelerator_memory_to_sram(accelerator_dram_address=GAMMA_DRAM_ADDR,
-                                    sram_address=gamma_sram_addr, element_size=N)
-        params_sram_addr += N * 2
-
-    if BETA_DRAM_ADDR is not None:
-        beta_sram_addr = params_sram_addr
-        ue.accelerator_memory_to_sram(accelerator_dram_address=BETA_DRAM_ADDR,
-                                    sram_address=beta_sram_addr, element_size=N)
-        params_sram_addr += N * 2
-
-    vector_A_sram_addr = 0x00000
-    vector_B_sram_addr = params_sram_addr
-    uram_b_remaining_elements = URAM_NEAR_FULL_ELEMENTS - (params_sram_addr - 0x80000) // 2
-    chunk_size = min(URAM_NEAR_FULL_ELEMENTS // N, uram_b_remaining_elements // N, M)
-    assert chunk_size >= 1 and chunk_size <= M
-
-    for i, m_take in ue.chunk_ranges(M, chunk_size):
-        chunk_elements = m_take * N
-        ue.accelerator_memory_to_sram(accelerator_dram_address=A_DRAM_ADDR + i * N * 2,
-                                    sram_address=vector_A_sram_addr, element_size=chunk_elements)
-        ue.accelerator_memory_to_sram(accelerator_dram_address=B_DRAM_ADDR + i * N * 2,
-                                    sram_address=vector_B_sram_addr, element_size=chunk_elements)
-        for j in range(m_take):
-            ue.eltwise_add_core(vector_A_sram_addr + j * N * 2, vector_B_sram_addr + j * N * 2,
-                                vector_A_sram_addr + j * N * 2, N)
-        ue.sram_to_accelerator_memory(sram_address=vector_A_sram_addr,
-                                        accelerator_dram_address=ADDOUTPUT_DRAM_ADDR + i * N * 2,
-                                        element_size=chunk_elements)
-        for j in range(m_take):
-            ue.layer_norm_core(vector_A_sram_addr + j * N * 2, vector_A_sram_addr + j * N * 2,
-                               N, zeros_sram_addr, gamma_sram_addr, beta_sram_addr)
-        ue.sram_to_accelerator_memory(sram_address=vector_A_sram_addr,
-                                        accelerator_dram_address=NORMOUTPUT_DRAM_ADDR + i * N * 2,
-                                        element_size=chunk_elements)
-
-    total_flops = M * N + 5 * M * N
-    if gamma_sram_addr is not None:
-        total_flops += M * N
-    if beta_sram_addr is not None:
-        total_flops += M * N
-    return total_flops
 def rope_hf_core_dram(ue, M: int, D: int,
                       X_DRAM_ADDR: int, OUTPUT_DRAM_ADDR: int,
                       COS_DRAM_ADDR: int, SIN_DRAM_ADDR: int,
@@ -579,14 +355,6 @@ def bf16_smart_permute_core(ue, dims, permute_indices, input_dram_addr, output_d
             remaining -= cur; out_addr += n_blocks * last_dim * bpe; i += cur
 
     return (2, output_shape)
-def store_identity_matrix(ue):
-    """Store identity matrix in DRAM once. Returns DRAM address."""
-    bpe = 2
-    size = UE_VECTOR_SIZE * UE_VECTOR_SIZE * bpe
-    addr = ue.get_params_dram_addr()
-    ue.dma_write(DMA_DEVICE_H2C, addr, torch.eye(UE_VECTOR_SIZE, dtype=torch.bfloat16), size)
-    ue.allocate_params_dram(size)
-    return addr
 def decode_flash_attention_core(ue, head_dim: int, kv_len: int,
                                  Q_DRAM_ADDR: int, K_DRAM_ADDR: int, V_DRAM_ADDR: int,
                                  OUTPUT_DRAM_ADDR: int, SCRATCH_DRAM_ADDR: int,
@@ -1004,40 +772,15 @@ def prefill_flash_attention_core(ue, head_dim: int, seq_len: int,
 # =============================================================================
 # GGUF generation — quantization helpers
 # =============================================================================
-# FP4 E2M1 lookup table: 16 values (codes 0-7 positive, 8-15 negative)
+# FP4 E2M1 lookup table: 16 values (codes 0-7 positive, 8-15 negative). Config-driven
+# so smolvlm2_config.json stays authoritative; passed into the shared quantizer.
 _FP4_E2M1_TABLE = torch.tensor(_SMOLVLM2_CFG["quantization"]["fp4_e2m1"]["table"], dtype=torch.bfloat16)
+
 def quantize_q4_64(tensor):
-    """INT4 quantization with 64-element blocks. Returns (packed_bytes, n_blocks)."""
-    data = tensor.flatten().cpu().float().numpy()
-    n_blocks = int(np.ceil(len(data) / 64))
-    padded = np.pad(data, (0, n_blocks * 64 - len(data)))
-    blocks = padded.reshape(n_blocks, 64)
-    scales = np.max(np.abs(blocks), axis=1)
-    scales[scales == 0] = 1.0
-    scales /= 7.0
-    quantized = np.clip(np.round(blocks / scales[:, None]), -8, 7).astype(np.int8)
-    pairs = (quantized.astype(np.uint8) & 0x0F).reshape(n_blocks, 32, 2)
-    packed = pairs[:, :, 0] | (pairs[:, :, 1] << 4)
-    scale_bytes = torch.tensor(scales, dtype=torch.float32).to(torch.bfloat16).view(torch.uint16).numpy()
-    return np.frombuffer(scale_bytes.tobytes() + packed.tobytes(), dtype=np.uint8), n_blocks
+    return _mlc_quantize_q4_64(tensor)
+
 def quantize_fp4_64(tensor):
-    """FP4 E2M1 quantization with 64-element blocks. Returns (packed_bytes, n_blocks)."""
-    x = tensor.to(torch.bfloat16).cpu().flatten()
-    n_blocks = int(np.ceil(x.numel() / 64))
-    if x.numel() % 64 != 0:
-        x = torch.nn.functional.pad(x, (0, n_blocks * 64 - x.numel()))
-    blocks = x.view(n_blocks, 64)
-    fp4_max = torch.tensor(6.0, dtype=torch.bfloat16)
-    scales = (blocks.abs().max(dim=1).values / fp4_max).clamp(min=1e-8).to(torch.bfloat16)
-    scaled = (blocks / scales[:, None]).to(torch.bfloat16)
-    # Nearest-neighbor lookup into FP4 table
-    codes = torch.argmin(torch.abs(scaled.unsqueeze(-1) - _FP4_E2M1_TABLE), dim=-1).to(torch.uint8)
-    codes_np = codes.numpy().flatten()
-    if len(codes_np) % 2 != 0:
-        codes_np = np.pad(codes_np, (0, 1))
-    packed = (codes_np[0::2] & 0x0F) | ((codes_np[1::2] & 0x0F) << 4)
-    scales_np = scales.view(torch.uint16).numpy()
-    return np.frombuffer(scales_np.tobytes() + packed.tobytes(), dtype=np.uint8), n_blocks
+    return _mlc_quantize_fp4_64(tensor, fp4_table=_FP4_E2M1_TABLE)
 # =============================================================================
 # Weight generation — name mapping, quantization dispatch, bin+json writers
 # =============================================================================
@@ -1553,7 +1296,7 @@ class SmolVLM2_UnifiedEngine(UnifiedEngine):
             h_in  = self.VIS_IO_A_DRAM if layer_idx % 2 == 0 else self.VIS_IO_B_DRAM
             h_out = self.VIS_IO_B_DRAM if layer_idx % 2 == 0 else self.VIS_IO_A_DRAM
             # LN1
-            layer_norm_core_dram(self, M=S, N=H, A_DRAM_ADDR=h_in, OUTPUT_DRAM_ADDR=self.VIS_LN_OUT_DRAM,
+            self.layer_norm_core_dram(M=S, N=H, A_DRAM_ADDR=h_in, OUTPUT_DRAM_ADDR=self.VIS_LN_OUT_DRAM,
                 GAMMA_DRAM_ADDR=la['ln1_weight'], BETA_DRAM_ADDR=la['ln1_bias'])
             # Q/K/V projections
             for proj, dst in [('q', self.VIS_Q_DRAM), ('k', self.VIS_K_DRAM), ('v', self.VIS_V_DRAM)]:
@@ -1589,7 +1332,7 @@ class SmolVLM2_UnifiedEngine(UnifiedEngine):
                 A_DRAM_ADDR=self.VIS_RESIDUAL_DRAM, B_DRAM_ADDR=self.VIS_MLP_OUT_DRAM, OUTPUT_DRAM_ADDR=h_out)
         # Post-layernorm
         final_vis = self.VIS_IO_A_DRAM if len(self.vis_layer_addrs) % 2 == 0 else self.VIS_IO_B_DRAM
-        layer_norm_core_dram(self, M=S, N=H, A_DRAM_ADDR=final_vis, OUTPUT_DRAM_ADDR=self.VIS_POST_LN_DRAM,
+        self.layer_norm_core_dram(M=S, N=H, A_DRAM_ADDR=final_vis, OUTPUT_DRAM_ADDR=self.VIS_POST_LN_DRAM,
             GAMMA_DRAM_ADDR=self.vis_post_ln_weight, BETA_DRAM_ADDR=self.vis_post_ln_bias)
         # Pixel shuffle: [1024,768] → [64,12288]
         bf16_smart_permute_core(self, dims=[8, 4, 8, 4, H], permute_indices=[0, 2, 1, 3, 4],
